make yaw bias estimator and simple 1D filter as individual HPP files
